Remove gameboard pckg
