+++ internal/app/tetris/gamestate_test.go
@@ -4,18 +4,17 @@ import (
 	"testing"
 
 	"github.com/Kaamkiya/gg/internal/app/tetris/color"
-	"github.com/Kaamkiya/gg/internal/app/tetris/gameboard"
 )
 
 func TestASingleLineIsRemoved(t *testing.T) {
 	gamestate := GameState{
 		nil,
 		nil,
-		gameboard.NewGameboard(color.Colors),
+		NewGameboard(color.Colors),
 	}
 
-	for i := range gameboard.Width {
-		gamestate.gameBoard.Grid[gameboard.Height-1][i] = color.Blue
+	for i := range Width {
+		gamestate.gameBoard.Grid[Height-1][i] = color.Blue
 	}
 
 	gamestate.handleCompletedLines(19, 19)
@@ -30,29 +29,29 @@ func TestMultipleLinesAreRemoved(t *testing.T) {
 	gamestate := GameState{
 		nil,
 		nil,
-		gameboard.NewGameboard(color.Colors),
+		NewGameboard(color.Colors),
 	}
 
-	for i := range gameboard.Width {
-		gamestate.gameBoard.Grid[gameboard.Height-1][i] = color.Blue
-		gamestate.gameBoard.Grid[gameboard.Height-3][i] = color.Blue
-		gamestate.gameBoard.Grid[gameboard.Height-4][i] = color.Blue
+	for i := range Width {
+		gamestate.gameBoard.Grid[Height-1][i] = color.Blue
+		gamestate.gameBoard.Grid[Height-3][i] = color.Blue
+		gamestate.gameBoard.Grid[Height-4][i] = color.Blue
 	}
 
-	gamestate.gameBoard.Grid[gameboard.Height-2][0] = color.Blue
-	gamestate.gameBoard.Grid[gameboard.Height-5][0] = color.Blue
+	gamestate.gameBoard.Grid[Height-2][0] = color.Blue
+	gamestate.gameBoard.Grid[Height-5][0] = color.Blue
 
 	gamestate.handleCompletedLines(16, 19)
 
-	if gamestate.gameBoard.Grid[gameboard.Height-1][0] != color.Blue && gamestate.gameBoard.Grid[gameboard.Height-1][1] != color.Black {
+	if gamestate.gameBoard.Grid[Height-1][0] != color.Blue && gamestate.gameBoard.Grid[Height-1][1] != color.Black {
 		t.Fatal("Second to last line didn't drop when last line was completed")
 	}
 
-	if gamestate.gameBoard.Grid[gameboard.Height-2][0] != color.Blue && gamestate.gameBoard.Grid[gameboard.Height-2][1] != color.Black {
+	if gamestate.gameBoard.Grid[Height-2][0] != color.Blue && gamestate.gameBoard.Grid[Height-2][1] != color.Black {
 		t.Fatal("Fifth to last line didn't drop when third to last line was completed")
 	}
 
-	if !gamestate.isLineEmpty(gameboard.Height - 3) {
+	if !gamestate.isLineEmpty(Height - 3) {
 		t.Fatal("Lines didn't move correctly when lines where completed")
 	}
